WIP : Menu Factory Reset
1. add factory reset for menu items.

+++ mobile/src/main/java/com/ctyeung/runyasso800/stateMachine/StateJog.kt
@@ -13,7 +13,7 @@ class StateJog : MotionState, Ijog, Isprint, Idone {
                 runViewModel: RunViewModel,
                 stepViewModel: StepViewModel):super(listener, actListener, runViewModel, stepViewModel)
     {
-        FINISH_DISTANCE = SharedPrefUtility.getDistance(SharedPrefUtility.keyJogDis)
+        FINISH_DISTANCE = SharedPrefUtility.getDistance(SharedPrefUtility.keyJogDis).toDouble()
     }
 
     /*
